Move webdriver client to tools/webdriver
The patch also ensures current WebDriver spec test behaviour doesn't
break.
