+++ webdriver/modal/alerts_quit_test.py
@@ -1,20 +1,14 @@
-# -*- mode: python; fill-column: 100; comment-column: 100; -*-
-
 import os
 import sys
 import unittest
 
-sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), os.path.pardir)))
+sys.path.insert(1, os.path.abspath(os.path.join(__file__, "../..")))
 import base_test
-
-from client.exceptions import NoSuchAlertException
-from client.wait import WebDriverWait
 
 
 class AlertsQuitTest(base_test.WebDriverBaseTest):
-
     def setUp(self):
-        self.wait = WebDriverWait(self.driver, 5, ignored_exceptions = [NoSuchAlertException])
+        self.wait = wait.WebDriverWait(self.driver, 5, ignored_exceptions=[exceptions.NoSuchAlertException])
         self.driver.get(self.webserver.where_is('modal/res/alerts.html'))
 
     def test_can_quit_when_an_alert_is_present(self):
@@ -24,6 +18,7 @@ class AlertsQuitTest(base_test.WebDriverBaseTest):
         with self.assertRaises(Exception):
             alert.accept()
         AlertsQuitTest.driver = None
+
 
 if __name__ == '__main__':
     unittest.main()
