+++ webdriver/timeouts/page_load_timeouts_tests.py
@@ -1,18 +1,15 @@
-# -*- mode: python; fill-column: 100; comment-column: 100; -*-
-
 import os
 import sys
 import unittest
-from selenium.common.exceptions import TimeoutException
 
-sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), os.path.pardir)))
+sys.path.insert(1, os.path.abspath(os.path.join(__file__, "../..")))
 import base_test
 
-class PageLoadTimeoutTest(base_test.WebDriverBaseTest):
 
+class PageLoadTimeoutTest(base_test.WebDriverBaseTest):
     def test_should_timeout_on_page_load_taking_too_long(self):
         self.driver.set_page_load_timeout(0.01)
-        with self.assertRaises(TimeoutException):
+        with self.assertRaises(exceptions.TimeoutException):
             self.load_page()
 
     def test_should_not_timeout_on_page_load(self):
@@ -22,6 +19,7 @@ class PageLoadTimeoutTest(base_test.WebDriverBaseTest):
 
     def load_page(self):
         self.driver.get(self.webserver.where_is('timeouts/res/page_load_timeouts_tests.html'))
+
 
 if __name__ == "__main__":
     unittest.main()
